+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to
 
 - 🐛(frontend) fix hand icon and queue position alignment and position #1119
 - 🩹(backend) add page_size to pagination for room endpoints #1131
+- 🐛(backend) refactor lobby throttling using participant id instead of IP #1129
 
 ## [1.10.0] - 2026-03-05